[fs] sunrpc: Add helpers to prevent socket create from racing
Message-id: <1447263055-22457-6-git-send-email-steved@redhat.com>
Patchwork-id: 126560
O-Subject: [RHEL7.3 fs PATCH 05/14] SUNRPC: Add helpers to prevent socket create from racing
Bugzilla: 1270038
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

The socket lock is currently held by the task that is requesting the
connection be established. While that is efficient in the case where
the connection happens quickly, it is racy in the case where it doesn't.
What we really want is for the connect helper to be able to block access
to the socket while it is being set up.

This patch does so by arranging to transfer the socket lock from the
task that is requesting the connect attempt, and then releasing that
lock once everything is done.
This scheme also gives us automatic protection against collisions with
the RPC close code, so we can kill the cancel_delayed_work_sync()
call in xs_close().

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream id: 718ba5b87343df303017585200ee182e937eabfc
Bugzilla: 1270038

Signed-off-by: Rafael Aquini <aquini@redhat.com>