[kernel] sched: Fix the group_capacity computation
Message-id: <1448462497-23895-13-git-send-email-jolsa@redhat.com>
Patchwork-id: 127614
O-Subject: [PATCH RHEL7 BZ1211784 12/28] sched/fair: Fix the group_capacity computation
Bugzilla: 1211784
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Daniel Bristot de Oliveira <bristot@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1211784
https://bugzilla.redhat.com/show_bug.cgi?id=1211784

upstream
========
commit c61037e905a5cb74c7d786c35ee2cdbab9ed63af
Author: Peter Zijlstra <peterz@infradead.org>
Date: Wed Aug 28 12:40:38 2013 +0200

description
===========
Do away with 'phantom' cores due to N*frac(smt_power) >= 1 by limiting
the capacity to the actual number of cores.

The assumption of 1 < smt_power < 2 is an actual requirement because
of what SMT is so this should work regardless of the SMT
implementation.

It can still be defeated by creative use of cpu hotplug, but if you're
one of those freaks, you get to live with it.

Signed-off-by: Rafael Aquini <aquini@redhat.com>