[kernel] sched: Change "has_capacity" to "has_free_capacity"
Message-id: <1448462497-23895-24-git-send-email-jolsa@redhat.com>
Patchwork-id: 127625
O-Subject: [PATCH RHEL7 BZ1211784 23/28] sched/fair: Change "has_capacity" to "has_free_capacity"
Bugzilla: 1211784
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Daniel Bristot de Oliveira <bristot@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1211784
https://bugzilla.redhat.com/show_bug.cgi?id=1211784

upstream
========
commit 1b6a7495d343fcfe22ff3a8285544bb8e40f1920
Author: Nicolas Pitre <nicolas.pitre@linaro.org>
Date: Mon May 26 18:19:35 2014 -0400

description
===========
The capacity of a CPU/group should be some intrinsic value that doesn't
change with task placement.  It is like a container which capacity is
stable regardless of the amount of liquid in it (its "utilization")...
unless the container itself is crushed that is, but that's another story.

Therefore let's rename "has_capacity" to "has_free_capacity" in order to
better convey the wanted meaning.

Signed-off-by: Rafael Aquini <aquini@redhat.com>