[fs] audit: implement audit by executable
Message-id: <9d2cdbe1961595f8e08c91b499304f0680da4291.1448476089.git.rgb@redhat.com>
Patchwork-id: 127682
O-Subject: [RHEL7 PATCH V3 25/26] audit: implement audit by executable
Bugzilla: 1135562
RH-Acked-by: Paul Moore <pmoore@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

This adds the ability audit the actions of a not-yet-running process.

This patch implements the ability to filter on the executable path.  Instead of
just hard coding the ino and dev of the executable we care about at the moment
the rule is inserted into the kernel, use the new audit_fsnotify
infrastructure to manage this dynamically.  This means that if the filename
does not yet exist but the containing directory does, or if the inode in
question is unlinked and creat'd (aka updated) the rule will just continue to
work.  If the containing directory is moved or deleted or the filesystem is
unmounted, the rule is deleted automatically.  A future enhancement would be to
have the rule survive across directory disruptions.

This is a heavily modified version of a patch originally submitted by Eric
Paris with some ideas from Peter Moody.

Cc: Peter Moody <peter@hda3.com>
Cc: Eric Paris <eparis@redhat.com>
Signed-off-by: Richard Guy Briggs <rgb@redhat.com>
[PM: minor whitespace clean to satisfy ./scripts/checkpatch]
Signed-off-by: Paul Moore <pmoore@redhat.com>
(cherry picked from commit 34d99af52ad40bd498ba66970579a5bc1fb1a3bc)
(fixed merge conflict with 51cc83f0 audit: add audit_backlog_wait_time
configuration option)

Signed-off-by: Richard Guy Briggs <rgb@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>