[fs] nfsd4: fix gss-proxy 4.1 mounts for some AD principals
Message-id: <1448750431-13603-6-git-send-email-bfields@redhat.com>
Patchwork-id: 127690
O-Subject: [Patch RHEL7 fs 5/5] nfsd4: fix gss-proxy 4.1 mounts for some AD principals
Bugzilla: 1283341
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

The principal name on a gss cred is used to setup the NFSv4.0 callback,
which has to have a client principal name to authenticate to.

That code wants the name to be in the form servicetype@hostname.
rpc.svcgssd passes down such names (and passes down no principal name at
all in the case the principal isn't a service principal).

gss-proxy always passes down the principal name, and passes it down in
the form servicetype/hostname@REALM.  So we've been munging the name
gss-proxy passes down into the format the NFSv4.0 callback code expects,
or throwing away the name if we can't.

Since the introduction of the MACH_CRED enforcement in NFSv4.1, we've
also been using the principal name to verify that certain operations are
done as the same principal as was used on the original EXCHANGE_ID call.

For that application, the original name passed down by gss-proxy is also
useful.

Lack of that name in some cases was causing some kerberized NFSv4.1
mount failures in an Active Directory environment.

This fix only works in the gss-proxy case.  The fix for legacy
rpc.svcgssd would be more involved, and rpc.svcgssd already has other
problems in the AD case.

Reported-and-tested-by: James Ralston <ralston@pobox.com>
Acked-by: Simo Sorce <simo@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>