[net] add sk_fullsock() helper
Message-id: <1452002044-19990-2-git-send-email-fwestpha@redhat.com>
Patchwork-id: 131639
O-Subject: [PATCH 1/4] [RHEL7.3 net] net: add sk_fullsock() helper
Bugzilla: 1295557
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Hannes Frederic Sowa <hannes@redhat.com>
RH-Acked-by: Sabrina Dubroca <sdubroca@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Conflicts: \
	include/net/sock.h

RHEL7 lacks TCPF_NEW_SYN_RECV state, there are also minor differences
in surrounding code ("extern" for instance).

This is only part of the series to ease future backports; in RHEL7
its enough to "sk_state != TCP_TIME_WAIT" to check if sk is a full socket
or not. upstream this helper has to be used instead.  So just put it
into RHEL7 so we don't have to mangle patches that use it.

commit 1d0ab253872cdd3d8e7913f59c266c7fd01771d0
Author: Eric Dumazet <edumazet@google.com>
Date:   Sun Mar 15 21:12:12 2015 -0700

    net: add sk_fullsock() helper

    We have many places where we want to check if a socket is
    not a timewait or request socket. Use a helper to avoid
    hard coding this.

    Signed-off-by: Eric Dumazet <edumazet@google.com>
    Signed-off-by: David S. Miller <davem@davemloft.net>

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>