[netdrv] igb: Teardown SR-IOV before unregister_netdev()
Message-id: <20160107210529.17767.66759.stgit@gimli.home>
Patchwork-id: 131743
O-Subject: [RHEL7.3 kernel PATCH 1/2] igb: Teardown SR-IOV before unregister_netdev()
Bugzilla: 1245562
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Bugzilla: 1245562
Upstream: c23d92b80e0b44d4c17085f0413e7574a7583615

When the .remove() callback for a PF is called, SR-IOV support for the
device is disabled, which requires unbinding and removing the VFs.
The VFs may be in-use either by the host kernel or userspace, such as
assigned to a VM through vfio-pci.  In this latter case, the VFs may
be removed either by shutting down the VM or hot-unplugging the
devices from the VM.  Unfortunately in the case of a Windows 2012 R2
guest, hot-unplug is broken due to the ordering of the PF driver
teardown.  Disabling SR-IOV prior to unregister_netdev() avoids this
issue.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Acked-by: Mitch Williams <mitch.a.williams@intel.com>
Tested-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>