[netdrv] ixgbe: Teardown SR-IOV before unregister_netdev()
Message-id: <20160107210535.17767.25360.stgit@gimli.home>
Patchwork-id: 131744
O-Subject: [RHEL7.3 kernel PATCH 2/2] ixgbe: Teardown SR-IOV before unregister_netdev()
Bugzilla: 1245562
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Bugzilla: 1245562
Upstream: 6b010e9b1f0a406d1d35202a694fa724a559bf77

When the .remove() callback for a PF is called, SR-IOV support for the
device is disabled, which requires unbinding and removing the VFs.
The VFs may be in-use either by the host kernel or userspace, such as
assigned to a VM through vfio-pci.  In this latter case, the VFs may
be removed either by shutting down the VM or hot-unplugging the
devices from the VM.  Unfortunately in the case of a Windows 2012 R2
guest, hot-unplug is broken due to the ordering of the PF driver
teardown.  Disabling SR-IOV prior to unregister_netdev() avoids this
issue.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Acked-by: Mitch Williams <mitch.a.williams@intel.com>
Tested-by: Krishneil Singh <krishneil.k.singh@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>