[fs] gfs2: Reintroduce a timeout in function gfs2_gl_hash_clear
Message-id: <2043434394.8684151.1452691354066.JavaMail.zimbra@redhat.com>
Patchwork-id: 131935
O-Subject: [RHEL7.3 fs PATCH] GFS2: Reintroduce a timeout in function gfs2_gl_hash_clear
Bugzilla: 1292582
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

I've already posted this patch for RHEL6, but this is the RHEL7 port.

This patch has been in many of my test kernels for the past several
months, but I did another kernel build just to be sure:

https://brewweb.devel.redhat.com/taskinfo?taskID=10332184

It was tested on system gfs-i24c-01.mpc.lab.eng.bos.redhat.com.

Patch description:

At some point in the past, we used to have a timeout when GFS2 was
unmounting, trying to clear out its glocks. If the timeout expires,
it would dump the remaining glocks to the kernel messages so that
developers can debug the problem. That timeout was eliminated,
probably by accident. This patch reintroduces it.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1292582
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>