[redhat] configs: Make NETPRIO_CGROUP a builtin
Message-id: <1453930870-15849-1-git-send-email-nhorman@redhat.com>
Patchwork-id: 133267
O-Subject: [RHEL7.3 PATCH v2] redhat: Make NETPRIO_CGROUP a builtin
Bugzilla: 1262204
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Intel requested that we make the netprio cgroup a builtin module.  As it
happens, systemd/dracut/something recently changed how cgroups are mounted,
converting the parent directory to a read-only filesystem, which means that if
you load the net_prio cgroup module after boot, you have to do some remounting
magic to get the new cgroup file system mounted at the same place as the others,
which is irritating.  I initially pushed back, but the argument was that it
should be easy make the netprio cgroup builtin.

Not so much.

Doing the obvious config change caused a _ton_ of ABI breaks.  And not just in
networking code where perhaps there was some minor semanic change in the
structures that are dependent on the state of CONFIG_NETPRIO_CGROUP.  Stuff in
the block layer and the like, really wierd stuff.

3 Days of staring at it led me to the problem.  Its the definition of the
enumeration of the cgroup subsys ids:

/*
 * Define the enumeration of all cgroup subsystems.
 *
 * We define ids for builtin subsystems and then modular ones.
 */
enum cgroup_subsys_id {
        CGROUP_BUILTIN_SUBSYS_COUNT,

        __CGROUP_SUBSYS_TEMP_PLACEHOLDER = CGROUP_BUILTIN_SUBSYS_COUNT - 1,

        CGROUP_SUBSYS_COUNT,
};

the sbsys array for a cgroup is defined by that CGROUP_SUBSYS_COUNT value, which
should be a static value, and it is, but not in the typical sense.  Because of
that intermediate value, the c preprocessor evaluates CGROUP_SUBSYS_COUNT to be
(10 - 1) + 2, which is what gets passed to the compiler.  while thats well and
good, by switching netprio to be builtin, it toggles its inclusion in this
enumeration to come after the PLACEHOLDER value, causing CGROUP_SUBSYS_COUNT to
be evaluated as (11 - 1) + 1.  I verified this by saving off all the
preprocessed files of a build and manually running them through genksyms.

While the math still works out to not change the size of any array defined by
CGROUP_SUBSYS_COUNT, since its the preprocessed, uncompiled code that gets used
by genksyms, this results in a change to the hash computed by any struct
containing an array defined by this ennumaration value.

So to fix it, I made this hot mess of a hack.  I changed removed the SUBSYS
macro use for the netprio id, and replaced it with a macro that I manually
enable so that the position of the id is the same as it was before, so
CGROUP_SUBSYS_COUNT doesn't change.

I could have done a GENKSYM around the whole enumeration, and replaced it with
another that simply defined CGROUP_SUBSYS_COUNT as (10 -1) + 2, but I was
concerned that it would mask problems if we ever added a new cgroup, so I
figured this was the better solution, messy though it was.

BZ: 1262204:
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9963260
Testing: Tested by myself manually in beaker

Change notes:

V2) Added comments in the cgroup header to explain what macro sorcery was afoot
as per Don Z's request.  Patch is otherwise unchanged

Signed-off-by: Rafael Aquini <aquini@redhat.com>