[kernel] perf: Fix running time accounting
Message-id: <1455060817-26852-27-git-send-email-jolsa@redhat.com>
Patchwork-id: 133816
O-Subject: [PATCH RHEL7.3 BZ1305882 026/360] perf: Fix running time accounting
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit 00a2916f7f82c348a2a94dbb572874173bc308a3
Author: Peter Zijlstra <peterz@infradead.org>
Date: Mon Jul 27 10:35:07 2015 +0200

description
===========
A recent fix to the shadow timestamp inadvertly broke the running time
accounting.

We must not update the running timestamp if we fail to schedule the
event, the event will not have ran. This can (and did) result in
negative total runtime because the stopped timestamp was before the
running timestamp (we 'started' but never stopped the event -- because
it never really started we didn't have to stop it either).

Fixes: 72f669c0086f ("perf: Update shadow timestamp before add event")
Signed-off-by: Rafael Aquini <aquini@redhat.com>