[tools] perf auxtrace: Add a heap for sorting AUX area tracing queues
Message-id: <1455060817-26852-54-git-send-email-jolsa@redhat.com>
Patchwork-id: 133843
O-Subject: [PATCH RHEL7.3 BZ1305882 053/360] perf auxtrace: Add a heap for sorting AUX area tracing queues
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit f939715586ea4f046eb52523ae25eb4d20b2a497
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Apr 9 18:53:52 2015 +0300

description
===========
In order to process AUX area tracing data in time order, the queue with
data with the lowest timestamp must be processed first.  Provide a heap
to keep track of which queue that is.

As with the queues, a decoder does not have to use the heap, but Intel
BTS and Intel PT will use it.

Signed-off-by: Rafael Aquini <aquini@redhat.com>