[tools] perf hists browser: Save pstack in the hist_browser
Message-id: <1455060817-26852-114-git-send-email-jolsa@redhat.com>
Patchwork-id: 133903
O-Subject: [PATCH RHEL7.3 BZ1305882 113/360] perf hists browser: Save pstack in the hist_browser
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit 01f00a1cd158ecd86d5f561ded271597d0550313
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Apr 22 16:18:16 2015 +0900

description
===========
The struct pstack is to save currently applied thread and/or dso filters
in the browser.  So it'd be better to keep in the struct hist_browser.
This is a preparation to later change.

Signed-off-by: Rafael Aquini <aquini@redhat.com>