[tools] perf hists browser: Save perf_session_env in the hist_browser
Message-id: <1455060817-26852-115-git-send-email-jolsa@redhat.com>
Patchwork-id: 133904
O-Subject: [PATCH RHEL7.3 BZ1305882 114/360] perf hists browser: Save perf_session_env in the hist_browser
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit b1a9ceef724341ce05b125d39abf9cfc7059b949
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Apr 22 16:18:17 2015 +0900

description
===========
The perf_session_env is to save system informantion at the recording
time to be refered in the hist browser.  So it'd be better to keep in
the struct hist_browser.  This is a preparation to later change.

Signed-off-by: Rafael Aquini <aquini@redhat.com>