[tools] perf probe: Fix an error when deleting probes successfully
Message-id: <1455060817-26852-208-git-send-email-jolsa@redhat.com>
Patchwork-id: 133997
O-Subject: [PATCH RHEL7.3 BZ1305882 207/360] perf probe: Fix an error when deleting probes successfully
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit dddc7ee32fa13efc66afa71ebd83bce545c8392a
Author: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Date: Wed May 27 17:37:25 2015 +0900

description
===========
Fix a bug in del_perf_probe_events() which returns an error (-ENOENT)
even if the probes are successfully deleted.

This happens only if the probes are on user-apps and not on kernel,
simply because it doesn't clear the previous error.

So, without this fix, we get an error even though events are being
successfully removed.

  ------
  # ./perf probe -x ./perf del_perf_probe_events
  Added new event:
    probe_perf:del_perf_probe_events (on del_perf_probe_events in ...

  You can now use it in all perf tools, such as:

          perf record -e probe_perf:del_perf_probe_events -aR sleep 1

  # ./perf probe -d \*:\*
  Removed event: probe_perf:del_perf_probe_events
    Error: Failed to delete events.
  ------

This fixes the above error.
  ------
  # ./perf probe -d \*:\*
  Removed event: probe_perf:del_perf_probe_events
  ------

Signed-off-by: Rafael Aquini <aquini@redhat.com>