[tools] perf tools: Remove newline char when reading event scale and unit
Message-id: <1455060817-26852-230-git-send-email-jolsa@redhat.com>
Patchwork-id: 134019
O-Subject: [PATCH RHEL7.3 BZ1305882 229/360] perf tools: Remove newline char when reading event scale and unit
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit 9ecae065f3295d7ea5e1c5492170a848ca255a46
Author: Madhavan Srinivasan <maddy@linux.vnet.ibm.com>
Date: Sun May 31 11:36:23 2015 +0530

description
===========
The <fd979c013207> commit intruduced the perf_event_sysfs_show function
to display the event_str value of an attr in kernel/event/core.c. But
the function returns the value with a newline char.

So, if a event also carries a event.unit file, when printing the counter
data perf tool formatting goes for a spin.

That is, because of the event unit, event name is printed in the newline
because of perf_event_sysfs_show returns with a newline char.

Now fixing perf core will break API, hencing proposing a fix in the perf tool.

[ Add spaces around operators ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>