[tools] perf probe: Fix to return error if no probe is added
Message-id: <1455060817-26852-268-git-send-email-jolsa@redhat.com>
Patchwork-id: 134057
O-Subject: [PATCH RHEL7.3 BZ1305882 267/360] perf probe: Fix to return error if no probe is added
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit b031220d520238075bd99513a420e65cf37866ad
Author: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Date: Tue Jun 16 20:50:55 2015 +0900

description
===========
Fix perf probe to return an error if no probe is added due to the given
probe point being on the blacklist.

To fix this problem, this moves the blacklist checking to right after
finding symbols/probe-points and marks them as skipped.

If all the symbols are skipped, "perf probe"  returns an error as it
fails to find the corresponding probe address.

E.g. currently if a blacklisted probe is given:

  # perf probe do_trap && echo 'succeed'
  Added new event:
  Warning: Skipped probing on blacklisted function: sync_regs
  succeed

No! It must fail! With this patch, it correctly fails:

  # perf probe do_trap && echo 'succeed'
  do_trap is blacklisted function, skip it.
  Probe point 'do_trap' not found.
    Error: Failed to add events.

Signed-off-by: Rafael Aquini <aquini@redhat.com>