[tools] perf report: Fix sort__sym_cmp to also compare end of symbol
Message-id: <1455060817-26852-284-git-send-email-jolsa@redhat.com>
Patchwork-id: 134073
O-Subject: [PATCH RHEL7.3 BZ1305882 283/360] perf report: Fix sort__sym_cmp to also compare end of symbol
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit c05676c06232e6459a6106ddf0d4e154ce6cd859
Author: Yannick Brosseau <scientist@fb.com>
Date: Wed Jun 17 16:41:10 2015 -0700

description
===========
When using a map file from a JIT, due to memory reuse, we can obtain
multiple symbols with the same start address but a different length.

The symbols__find does check for the end so not doing it in
sort__sym_cmp was causing the hist_entry in the annotate part of a
report to match to the wrong entry, causing a fatal error.

Signed-off-by: Rafael Aquini <aquini@redhat.com>