[tools] perf tools: Allow events with dot
Message-id: <1455060817-26852-294-git-send-email-jolsa@redhat.com>
Patchwork-id: 134083
O-Subject: [PATCH RHEL7.3 BZ1305882 293/360] perf tools: Allow events with dot
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit 5b021ddf81b0bcd17d88186b605f53a2ca452d3a
Author: Andi Kleen <ak@linux.intel.com>
Date: Wed Jun 10 00:25:11 2015 -0700

description
===========
The Intel events use a dot to separate event name and unit mask.  Allow
dot in names in the scanner, and remove special handling of dot as EOF.
Also remove the hack in jevents to replace dot with underscore. This way
dotted events can be specified directly by the user.

I'm not fully sure this change to the scanner is correct (what was the
dot special case good for?), but I haven't found anything that breaks
with it so far at least.

Signed-off-by: Rafael Aquini <aquini@redhat.com>