[tools] perf tools: Make fork event processing more resilient
Message-id: <1455060817-26852-359-git-send-email-jolsa@redhat.com>
Patchwork-id: 134148
O-Subject: [PATCH RHEL7.3 BZ1305882 358/360] perf tools: Make fork event processing more resilient
Bugzilla: 1305882
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1305882
https://bugzilla.redhat.com/show_bug.cgi?id=1305882

upstream
========
commit 5cb73340d92a716fd2776700742c3558206ae298
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Aug 19 17:29:20 2015 +0300

description
===========
When processing a fork event, the tools lookup the parent thread by its
tid.  In a couple of cases, it is possible for that thread to have the
wrong pid.

That can happen if the data is being processed out of order, or if the
(fork) event that would have removed the erroneous thread was lost.

Assume the latter case, print a dump message, remove the erroneous
thread, create a new one with the correct pid, and keep going.

Signed-off-by: Rafael Aquini <aquini@redhat.com>