[tools] perf evlist: Make perf_evlist__set_filter use perf_evsel__set_filter
Message-id: <1455542356-29341-37-git-send-email-jolsa@redhat.com>
Patchwork-id: 134416
O-Subject: [PATCH RHEL7.3 BZ1306834 036/237] perf evlist: Make perf_evlist__set_filter use perf_evsel__set_filter
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit 94ad89bc8a61b79c416c180c51b3e61ac16f6124
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Jul 3 17:42:03 2015 -0300

description
===========
Instead of calling perf_evsel__apply_filter straight away, so that
we can, in the next patches, expand the filter with more conditions
before actually calling the ioctl to pass the end result filter to
the kernel.

Now we need to call perf_evlist__apply_filters() after the filter
is completely setup, i.e. do the ioctl calls.

The perf_evlist__apply_filters() method was already in place, because
that is the model for the other tools that receives filters in the
command line: go on setting then in the evsel->filter and only at
the end, after parsing the whole command line, apply them.

We get, as a bonus, a more expressive message that states which
event, if any, failed to have the filter applied to, with an
error message stating what happened.

Signed-off-by: Rafael Aquini <aquini@redhat.com>