[tools] perf annotate: Compute IPC and basic block cycles
Message-id: <1455542356-29341-109-git-send-email-jolsa@redhat.com>
Patchwork-id: 134522
O-Subject: [PATCH RHEL7.3 BZ1306834 108/237] perf annotate: Compute IPC and basic block cycles
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit 30e863bb6f708c0abd422fbb0e6b295f5ee6407b
Author: Andi Kleen <ak@linux.intel.com>
Date: Sat Jul 18 08:24:50 2015 -0700

description
===========
Compute the IPC and the basic block cycles for the annotate display.

IPC is computed by counting the instructions, and then dividing the
accounted cycles by that count.

The actual IPC computation can only be done at annotate time, because we
need to parse the objdump output first to know the number of
instructions in the basic block.

The cycles/IPC are also put into the perf function annotation so that
the display code can show them.

Again basic block overlaps are not handled, with the longest winning,
but there are some heuristics to hide the IPC when the longest is not
the most common.

v2: Compute IPC correctly.
Signed-off-by: Rafael Aquini <aquini@redhat.com>