[tools] perf tools: Validate config term maximum value
Message-id: <1455542356-29341-117-git-send-email-jolsa@redhat.com>
Patchwork-id: 134480
O-Subject: [PATCH RHEL7.3 BZ1306834 116/237] perf tools: Validate config term maximum value
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit 0efe6b67690b6546daa0d2f34a17eb3ca46c9dea
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Jul 17 19:33:50 2015 +0300

description
===========
Currently the value of a PMU config term is silently truncated if it is
too big. This is an impediment to validating the value for other
criteria later on i.e.  the user provides an invalid value that gets
truncated to a valid one.

The maximum value validation is only done for the parser where the error
is passed back to the user. In other cases the silent truncation
continues so as not to affect tools that perhaps rely on it.

Signed-off-by: Rafael Aquini <aquini@redhat.com>