[tools] perf hists: Update the column width for the "srcline" sort key
Message-id: <1455542356-29341-128-git-send-email-jolsa@redhat.com>
Patchwork-id: 134497
O-Subject: [PATCH RHEL7.3 BZ1306834 127/237] perf hists: Update the column width for the "srcline" sort key
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit e8e6d37e73e6b950c891c780745460b87f4755b6
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Aug 10 16:53:54 2015 -0300

description
===========
When we introduce a new sort key, we need to update the
hists__calc_col_len() function accordingly, otherwise the width
will be limited to strlen(header).

We can't update it when obtaining a line value for a column (for
instance, in sort__srcline_cmp()), because we reset it all when doing a
resort (see hists__output_recalc_col_len()), so we need to, from what is
in the hist_entry fields, set each of the column widths.

Fixes: 409a8be61560 ("perf tools: Add sort by src line/number")
Signed-off-by: Rafael Aquini <aquini@redhat.com>