[tools] perf tools: /proc/kcore requires CAP_SYS_RAWIO message too noisy
Message-id: <1455542356-29341-160-git-send-email-jolsa@redhat.com>
Patchwork-id: 134531
O-Subject: [PATCH RHEL7.3 BZ1306834 159/237] perf tools: /proc/kcore requires CAP_SYS_RAWIO message too noisy
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit 133de940435493266acd4100b0dae8ba7eaa7c71
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Aug 20 13:07:40 2015 +0300

description
===========
The "/proc/kcore requires CAP_SYS_RAWIO" message comes up all the time
for 'perf script' if vmlinux is not found and the user isn't root, even
when the kernel is not being traced and even though the message is only
really relevant for annotation.

Change it to pr_debug and instead put a note in the message displayed if
annotation is not possible.

Also, the file being accessed might not be /proc/kcore.  Tools can be
directed to a different location using the --kallsyms option in which
case kcore is expected to be in the same directory.  Adjust the message
so it is not misleading in that case.

Signed-off-by: Rafael Aquini <aquini@redhat.com>