[tools] perf tools: Fix Intel PT timestamp handling
Message-id: <1455542356-29341-161-git-send-email-jolsa@redhat.com>
Patchwork-id: 134536
O-Subject: [PATCH RHEL7.3 BZ1306834 160/237] perf tools: Fix Intel PT timestamp handling
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit 81cd60cc29a9c3e92ad6ca167a4764b2c2d2fc04
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Aug 20 11:51:32 2015 +0300

description
===========
Events that don't sample the timestamp have a timestamp value of -1.

Intel PT processing wasn't taking that into account.

This is particularly noticeable with Intel BTS because timestamps are
not requested by default.

Then, if the conversion of -1 to TSC results in a small number, the
processing is unaffected.

However if the conversion results in a big number, then the data is
processed prematurely before relevant sideband data like mmap events,
which in turn results in samples with unknown dsos.

Commiter note:

Since BTS wasn't upstream, I split the patch to fold the BTS part with
the patch introducing it, to avoid having this bug in the commit
history. PT was already upstream, so this patch contains that part.

Signed-off-by: Rafael Aquini <aquini@redhat.com>