[tools] perf tools: Initialize reference counts in map__clone()
Message-id: <1455542356-29341-166-git-send-email-jolsa@redhat.com>
Patchwork-id: 134547
O-Subject: [PATCH RHEL7.3 BZ1306834 165/237] perf tools: Initialize reference counts in map__clone()
Bugzilla: 1306834
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1306834
https://bugzilla.redhat.com/show_bug.cgi?id=1306834

upstream
========
commit 66671d001aeb525b9101e0ccb9062627539de555
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Aug 18 15:19:50 2015 -0300

description
===========
Map clone was written before we introduced reference counts for
maps and dsos, so all that was needed was just a copy and then we
would insert it into the new map_groups instance.

Fix it by, after copying, initializing the map->refcnt, grabbing
a struct dso refcount and resetting pointers that may be used
to determine if a map, when deleted, is in a rb_tree.

Signed-off-by: Rafael Aquini <aquini@redhat.com>